+++ ComputerGraphics/graphics/geometry/TorusGeometry.java
@@ -14,13 +14,13 @@ public class TorusGeometry extends SurfaceGeometry
         super(
             (u, v) ->
             {
-                return new Vector( (1 + 0.25 * Math.cos(v)) * Math.cos(u),
-                                   (1 + 0.25 * Math.cos(v)) * Math.sin(u),
-                                    0.25 * Math.sin(v) );
+                return new Vector( (1 + 0.25f * (float)Math.cos(v)) * (float)Math.cos(u),
+                                   (1 + 0.25f * (float)Math.cos(v)) * (float)Math.sin(u),
+                                    0.25f * (float)Math.sin(v) );
                                     
             },
             
-            0, 2 * Math.PI, 20,
-            0, 2 * Math.PI, 20 );
+            0, 2 * (float)Math.PI, 20,
+            0, 2 * (float)Math.PI, 20 );
     }
 }
